large change, removed selector as core-concept, merged into environment-concept

+++ user/entity/markov2in1out.h
@@ -14,12 +14,12 @@ class markov2in1out {
   size_t hidden_ = 4;
   life::encoding genome_ = life::generate(100);
 
-  std::vector<size_t> buffer_ =
-      std::vector<size_t>(input_ + output_ + hidden_, 0u);
+  std::vector<double> buffer_ =
+      std::vector<double>(input_ + output_ + hidden_, 0u);
 
   struct gate {
     size_t in_1_, in_2_, out_;
-    std::array<size_t, 4> logic_;
+    std::array<long, 4> logic_;
   };
 
   std::vector<gate> gates_;
@@ -43,7 +43,7 @@ public:
     output_ = con["outputs"];
     hidden_ = con["hiddens"];
     life::encoding genome_ = life::generate(100);
-    buffer_ = std::vector<size_t>(input_ + output_ + hidden_, 0u);
+    buffer_ = std::vector<double>(input_ + output_ + hidden_, 0u);
     seed_gates(4);
   }
